requirements.txt
streamlit==1.35.0
qrcode[pil]==7.4.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
pyzbar==0.1.10
Pillow==10.3.0
pandas==2.2.2